Add tester.py and testHelpers.py

+++ e2e/FibonacciTest/gen.py
@@ -3,7 +3,7 @@ import random
 import os
 
 input_dir = r'Input'
-output_dir = r'Output'
+expected_dir = r'Expected'
 global_file_counter = 0
 random.seed()
 
@@ -26,8 +26,9 @@ def fibonacci_of(n):
 
 
 def remove_and_create_folders():
-    for directory in [input_dir, output_dir]:
-        shutil.rmtree(directory)
+    for directory in [input_dir, expected_dir]:
+        if os.path.exists(directory):
+            shutil.rmtree(directory)
         os.makedirs(directory)
 
 
@@ -40,12 +41,12 @@ def create_file(n=None):
     file_number = global_file_counter
     global_file_counter += 1
     input_file = os.path.join(input_dir, str(file_number)+ '.in')
-    output_file = os.path.join(output_dir, str(file_number) + '.out')
+    output_file = os.path.join(expected_dir, str(file_number) + '.out')
 
-    with open(input_file, 'a') as input:
+    with open(input_file, 'w') as input:
         print(n, file=input)
     
-    with open(output_file, 'a') as output:
+    with open(output_file, 'w') as output:
         print(fibonacci_of(n), file=output)
 
 def create_files(num_files = 10):
