+++ e2e/IsPrimeTest/gen.py
@@ -3,7 +3,7 @@ import random
 import os
 
 input_dir = r'Input'
-output_dir = r'Output'
+expected_dir = r'Expected'
 global_file_counter = 0
 random.seed()
 
@@ -20,8 +20,9 @@ def is_prime(n):
     return True
 
 def remove_and_create_folders():
-    for directory in [input_dir, output_dir]:
-        shutil.rmtree(directory)
+    for directory in [input_dir, expected_dir]:
+        if os.path.exists(directory):
+            shutil.rmtree(directory)
         os.makedirs(directory)
 
 def create_file(n=None):
@@ -33,7 +34,7 @@ def create_file(n=None):
     file_number = global_file_counter
     global_file_counter += 1
     input_file = os.path.join(input_dir, str(file_number)+ '.in')
-    output_file = os.path.join(output_dir, str(file_number) + '.out')
+    output_file = os.path.join(expected_dir, str(file_number) + '.out')
 
     with open(input_file, 'a') as input:
         print(n, file=input)
