+++ e2e/SumTest/gen.py
@@ -3,13 +3,14 @@ import random
 import os
 
 input_dir = r'Input'
-output_dir = r'Output'
+expected_dir = r'Expected'
 global_file_counter = 0
 random.seed()
 
 def remove_and_create_folders():
-    for directory in [input_dir, output_dir]:
-        shutil.rmtree(directory)
+    for directory in [input_dir, expected_dir]:
+        if os.path.exists(directory):
+            shutil.rmtree(directory)
         os.makedirs(directory)
 
 
@@ -19,7 +20,7 @@ def create_files(num_files = 5, numbers_count=1000, min_number = 1, max_number =
         file_number = global_file_counter + file_idx
         global_file_counter += 1
         input_file = os.path.join(input_dir, str(file_number)+ '.in')
-        output_file = os.path.join(output_dir, str(file_number) + '.out')
+        output_file = os.path.join(expected_dir, str(file_number) + '.out')
 
         data=[random.randrange(min_number, max_number) for _ in range(numbers_count)]
